refactor recalculateAndApplyInvoiceQuantity
